LoginController and AuthController are now merged
