+++ src/main/java/com/example/NetProjectBackend/confuguration/WebSecurityConfiguration.java
@@ -60,6 +60,7 @@ public class WebSecurityConfiguration extends WebSecurityConfigurerAdapter {
                 .antMatchers("/auth/**").permitAll()
                 .antMatchers("/recovery").permitAll()
                 .antMatchers("/hello/**").permitAll()
+                .antMatchers("/code").permitAll()
                 .anyRequest().authenticated();
         http.addFilterBefore(authenticationJwtTokenFilter(), UsernamePasswordAuthenticationFilter.class);
     }
